Выполнил ДЗ Модуляция
